+++ examples/use_rust/nrf52840_ble_split/src/central.rs
@@ -23,9 +23,9 @@ use rand_chacha::ChaCha12Rng;
 use rand_core::SeedableRng;
 use rmk::ble::trouble::build_ble_stack;
 use rmk::channel::EVENT_CHANNEL;
-use rmk::config::{
-    BehaviorConfig, BleBatteryConfig, ControllerConfig, KeyboardUsbConfig, RmkConfig, StorageConfig, VialConfig,
-};
+use rmk::config::{BehaviorConfig, BleBatteryConfig, KeyboardUsbConfig, RmkConfig, StorageConfig, VialConfig};
+use rmk::controller::EventController as _;
+use rmk::controller::led_indicator::{KeyboardIndicator, KeyboardIndicatorController};
 use rmk::debounce::default_debouncer::DefaultDebouncer;
 use rmk::futures::future::{join, join4};
 use rmk::input_device::Runnable;
@@ -33,7 +33,6 @@ use rmk::input_device::adc::{AnalogEventType, NrfAdc};
 use rmk::input_device::battery::BatteryProcessor;
 use rmk::input_device::rotary_encoder::RotaryEncoder;
 use rmk::keyboard::Keyboard;
-use rmk::light::LightController;
 use rmk::split::ble::central::read_peripheral_addresses;
 use rmk::split::central::{CentralMatrix, run_peripheral_manager};
 use rmk::{HostResources, initialize_encoder_keymap_and_storage, run_devices, run_processor_chain, run_rmk};
@@ -210,11 +209,7 @@ async fn main(spawner: Spawner) {
     // Read peripheral address from storage
     let peripheral_addrs = read_peripheral_addresses::<1, _, 8, 7, 4, 2>(&mut storage).await;
 
-    // Initialize the light controller
-    let mut light_controller: LightController<Output> = LightController::new(ControllerConfig::default().light_config);
-
     // Initialize the encoder processor
-
     let mut adc_device = NrfAdc::new(
         saadc,
         [AnalogEventType::Battery],
@@ -222,6 +217,17 @@ async fn main(spawner: Spawner) {
         None,
     );
     let mut batt_proc = BatteryProcessor::new(2000, 2806, &keymap);
+
+    // Initialize the controllers
+    let mut capslock_led = KeyboardIndicatorController::new(
+        Output::new(
+            p.P0_00,
+            embassy_nrf::gpio::Level::Low,
+            embassy_nrf::gpio::OutputDrive::Standard,
+        ),
+        false,
+        KeyboardIndicator::CapsLock,
+    );
 
     // Start
     join4(
@@ -231,10 +237,10 @@ async fn main(spawner: Spawner) {
         run_processor_chain! {
             EVENT_CHANNEL => [batt_proc],
         },
-        keyboard.run(),
+        join(keyboard.run(), capslock_led.event_loop()),
         join(
             run_peripheral_manager::<4, 7, 4, 0, _>(0, peripheral_addrs[0], &stack),
-            run_rmk(&keymap, driver, &stack, &mut storage, &mut light_controller, rmk_config),
+            run_rmk(&keymap, driver, &stack, &mut storage, rmk_config),
         ),
     )
     .await;