+++ rmk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rewrite led indicator, use controller system
+
 ## [0.7.8] - 2025-07-23
 
 ### Added